test: combine mcp tools and server defs
