--- tests/agents/durableagent/test_mcp_math_tools.py
@@ -1,28 +1,0 @@
-#
-# Copyright 2026 The Dapr Authors
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#     http://www.apache.org/licenses/LICENSE-2.0
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-
-from mcp.server.fastmcp import FastMCP
-
-mcp = FastMCP("MathServer")
-
-
-@mcp.tool()
-async def add(a: int, b: int) -> int:
-    """Add two numbers and return the result."""
-    return a + b
-
-
-@mcp.tool()
-async def multiply(a: int, b: int) -> int:
-    """Multiply two numbers and return the result."""
-    return a * b
